requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
geopandas>=0.14.0
fiona>=1.9.0
shapely>=2.0.0
pandas>=2.0.0
pyyaml>=6.0
